Fix bean accessors being used for field names
When using `@JsonProperty` on a bean property (i.e. a field with
accessor methods), if you do not specify an explicit name for the
property, the `ScalaAnnotationIntrospector` was incorrectly generating
properties for the accessor methods.

In the following example:

```
class Bean {
  private var value: Int = 0

  @JsonProperty
  def getValue: Int = value

  @JsonProperty
  def setValue(value: Int): Unit = { this.value = value }
}
```

A mapper configured with the `ScalaAnnotationIntrospector` would
serialize this to:

```
{"value":0,"getValue":0}
```

This most notably impacts interop with Java-defined classes, as they
often follow the pattern of Java Beans, where the accessor methods are
annotated with `@JsonProperty`.

Credit for the fix goes to @clintmiller1
(https://github.com/FasterXML/jackson-module-scala/issues/224#issuecomment-249909564).

I've included a test that verifies the problem, and the solution.